Pull in a workaround from sbruno@ to fix crt init functions.
Add a .local to force these functions to be treated as local.  In practice,
frame_dummy should be treated as local and the others as global, but
clang defaults to treating them all as global even though frame_dummy is
marked static.  This results in a CALL16 reloc against the private
frame_dummy symbol which fails to link.